+++ src/utils/generateDisplayName.js
@@ -4,5 +4,6 @@ import isTag from './isTag';
 import type { Target } from '../types';
 
 export default function generateDisplayName(target: Target): string {
+  // $FlowFixMe
   return isTag(target) ? `styled.${target}` : `Styled(${getComponentName(target)})`;
 }